host/trxcon: introduce application state/event manager
Since the application becomes more complex, there are some ways to
control all it's components and make them work together. One of them
is to use the finite-state machine. Currently there are two states:

- APP_STATE_IDLE - we have one active L1CTL connection;
- APP_STATE_MANAGED - there is no L1CTL connection, wait for a new one.

But instead of switching the application state directly, I prefer to
use the event-based switching model. The idea is to poke the event
handler (manager) every time, when something happens, providing an
event type. The handler, in its turn, will change the application state
depending on a current state, type of incoming event and optionally
depending on some other conditions.